To main from dev

+++ TicketsAPI/Repository/TicketRepository.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using TicketsAPI.Entities;
@@ -20,11 +21,12 @@ namespace TicketsAPI.Repository
                 .Tickets
                 .ToListAsync(cancelationToken);
         }
-        public async Task<Ticket> GetById(int id, CancellationToken cancelationToken)
+        public async Task<List<Ticket>> GetTicketListByUserId(int userId, CancellationToken cancelationToken)
         {
             return await dbContext
                 .Tickets
-                .FirstOrDefaultAsync(e => e.Id == id);
+                .Where( t => t.UserId == userId)
+                .ToListAsync();
         }
         public async Task AddTicketWithUserToEvent(Ticket ticket, CancellationToken cancelationToken)
         {
@@ -39,9 +41,15 @@ namespace TicketsAPI.Repository
             var ticket = await dbContext
                 .Tickets
                 .FirstOrDefaultAsync(e => e.Id == id);
+            var user = await dbContext
+                .Users
+                .FirstOrDefaultAsync(e => e.Id == ticket.UserId);
             if (ticket == null)
                 return false;
-
+            if (user == null)
+                return false;
+            user.Wallet += ticket.Price;
+            dbContext.Users.Update(user);
             dbContext.Tickets.Remove(ticket);
             await dbContext.SaveChangesAsync(cancelationToken);
             return true;
